Move layouts from widgets to models
