+++ bokeh/models/__init__.py
@@ -9,6 +9,7 @@ from .component import *
 from .formatters import *
 from .glyphs import *
 from .grids import *
+from .layouts import *
 from .images import *
 from .map_plots import *
 from .markers import *
